Add type_casts method
This method returns a row used for typecasting the values that will be
updated. This is needed because many types don't have a specific
literal syntax and are instead using the string literal syntax.
This will cause type mismatches because there's not context, which is
otherwise present for regular inserts, for the values to infer the
types from.